+++ oalg-abg/src/OAlg/AbelianGroup/Liftable.hs
@@ -50,10 +50,44 @@ import OAlg.Entity.Diagram
 import OAlg.Entity.Matrix
 import OAlg.Entity.Sequence.PSequence
 
+import OAlg.Limes.Definition
+import OAlg.Limes.KernelsAndCokernels
+
 import OAlg.AbelianGroup.Definition
 import OAlg.AbelianGroup.Free.SmithNormalForm
 import OAlg.AbelianGroup.Euclid
 
+--------------------------------------------------------------------------------
+-- abhCokerLft -
+
+-- | the associated liftable, i.e. every cokernel in 'AbHom' has the @'Liftable' 'From'@ property.
+--
+-- __Property__ Let @coker@ be in @'Cokernel' 'N1' 'AbHom@, than holds:
+--
+-- (1) @c '==' l@ where @c = 'cokernelFactor' ('universalCone' coker)@ and
+-- @'LiftableFrom' c _ = 'abhCokerLft' coker@.
+--
+-- @
+--         c         a
+--   * <------- * <------*
+--    ^         ^
+--     \       /
+--   f  \     / f' 
+--       \   /
+--        \ /
+--         *
+-- @
+--
+-- where @c@ is the cokernel of @a@ and @f'@ the lifted @f@. 
+abhCokerLft :: Attestable n => Cokernel N1 AbHom -> Liftable From (Free n) AbHom
+abhCokerLft coker = LiftableFrom c l where
+  c = cokernelFactor $ universalCone coker
+  l :: Slice From (Free n) AbHom -> Slice From (Free n) AbHom
+  l f | end c /= end (slice f) = throw NotLiftable
+      | otherwise              = error "nyi"
+
+
+{-
 --------------------------------------------------------------------------------
 -- abhLift -
 
@@ -107,7 +141,8 @@ prpAbhLift = Prp "AbhLift" :<=>:
                   in do
                     x <- xMatrixZ (d^k' :> d^n')
                     let x' = p * zabh x in return (a,someNatural k',a * x')
-                   
+-}
+
 --------------------------------------------------------------------------------
 -- zMatrixLift -